% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Shiny visualization app for Poisson auto-models}
\usage{
launchApp(sz = 15, M = 10)
}
\arguments{
\item{sz}{size of the regular lattice}

\item{M}{overall mean of the covariate}
}
\value{
shiny app
}
\description{
Shiny visualization app for Poisson auto-models
}
\examples{
\dontrun{
launchApp()
}
}

