% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrgrid_4nbrs.R
\name{sqrgrid.4nbrs}
\alias{sqrgrid.4nbrs}
\title{Program Computes neighbors for 4 nearest neighbor structure on a square regular lattice.}
\usage{
sqrgrid.4nbrs(k)
}
\arguments{
\item{k}{size of the regular lattice}
}
\value{
Matrix with neighbors information
}
\description{
Locations are assumed to be 1, . . .,k in row 1, k+1, . . .,2k in row two, etc.
}

